6-ое Задание

+++ 05.05.16/HomeWork_BankClient/Bank.cs
@@ -8,7 +8,7 @@ namespace HomeWork_BankClient
 {
     class Bank
     {
-        public void Transaction (BaseAccount from, BaseAccount to, double sum)
+        public void Transaction (BaseAccount from, BaseAccount to, int sum)
         {
             from.WithdrawMoneyFromAccount(sum);
             to.DepositAccount(sum);
